tests/test_visual_smoke.py
import pytest
from playwright.sync_api import Page

from tests.visual_expect import expect

DESKTOP_VIEWPORT = {"width": 1280, "height": 800}
MOBILE_VIEWPORT = {"width": 390, "height": 844}



def _open_title(page: Page, game_url: str, viewport: dict[str, int]) -> None:
    page.set_viewport_size(viewport)
    page.goto(game_url, wait_until="domcontentloaded")
    page.wait_for_selector("#passages .passage")



def _click_first_choice(page: Page) -> None:
    passage = page.locator("#passages .passage").last
    first_choice = passage.locator("a.link-internal").first
    first_choice.click()
    page.wait_for_timeout(250)



@pytest.mark.screenshot
def test_title_desktop(page: Page, game_url: str, baseline_dir, failure_dir, update_snapshots) -> None:
    _open_title(page, game_url, DESKTOP_VIEWPORT)
    expect(page, baseline_dir, failure_dir, update_snapshots).to_have_screenshot(
        "title.desktop.png", threshold=0.1
    )



@pytest.mark.screenshot
def test_title_mobile(page: Page, game_url: str, baseline_dir, failure_dir, update_snapshots) -> None:
    _open_title(page, game_url, MOBILE_VIEWPORT)
    expect(page, baseline_dir, failure_dir, update_snapshots).to_have_screenshot(
        "title.mobile.png", threshold=0.1
    )



@pytest.mark.screenshot
def test_hud_desktop(page: Page, game_url: str, baseline_dir, failure_dir, update_snapshots) -> None:
    _open_title(page, game_url, DESKTOP_VIEWPORT)
    _click_first_choice(page)
    expect(page, baseline_dir, failure_dir, update_snapshots).to_have_screenshot(
        "hud.desktop.png", threshold=0.1
    )



@pytest.mark.screenshot
def test_hud_mobile(page: Page, game_url: str, baseline_dir, failure_dir, update_snapshots) -> None:
    _open_title(page, game_url, MOBILE_VIEWPORT)
    _click_first_choice(page)
    expect(page, baseline_dir, failure_dir, update_snapshots).to_have_screenshot(
        "hud.mobile.png", threshold=0.1
    )



@pytest.mark.screenshot
def test_ending_desktop(page: Page, game_url: str, baseline_dir, failure_dir, update_snapshots) -> None:
    _open_title(page, game_url, DESKTOP_VIEWPORT)
    _click_first_choice(page)
    _click_first_choice(page)

    # No ending passage is reachable within three reliable clicks from Title in current flow,
    # so this captures the two-click fallback state (Hallway).
    expect(page, baseline_dir, failure_dir, update_snapshots).to_have_screenshot(
        "ending.desktop.png", threshold=0.1
    )
