requirements-dev.txt
pytest
pytest-playwright
pillow
